+++ src/main/java/com/buuz135/industrial/tile/agriculture/AnimalResourceHarvesterTile.java
@@ -98,6 +98,6 @@ public class AnimalResourceHarvesterTile extends WorkingAreaElectricMachine {
 
     @Override
     protected void processFluidItems(ItemStackHandler fluidItems) {
-        ItemStackUtils.processFluidItems(fluidItems, tank);
+        ItemStackUtils.fillItemFromTank(fluidItems, tank);
     }
 }