+++ src/main/java/com/buuz135/industrial/tile/generator/LavaFabricatorTile.java
@@ -44,6 +44,6 @@ public class LavaFabricatorTile extends CustomElectricMachine {
 
     @Override
     protected void processFluidItems(ItemStackHandler fluidItems) {
-        ItemStackUtils.processFluidItems(fluidItems, tank);
+        ItemStackUtils.fillItemFromTank(fluidItems, tank);
     }
 }