+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 + Insertion and extraction conveyor upgrades can now extract from entites
 + Right clicking conveyors with plastic will prevent items from being picked up
 + Changed how the book entries are displayed when hovered
++ Fixed cells only being processed 1 by 1
 
 # 1.10.0
 + Added 6 conveyor upgrades that can be added to the conveyor (Extraction, Insertion, Splitting, Bouncing, Dropping and Blinking) (Coded/Buuz135)