+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# 1.10.2
++ Fluid interaction has potion effects now
+
 # 1.10.1
 + If the black hole controller is lock the black hole units won't loose the filter
 + Insertion and extraction conveyor upgrades can now extract from entites